backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
face-recognition==1.3.0
dlib==19.22.99
numpy==1.24.3
Pillow==10.0.0
pymongo==4.5.0
python-dateutil==2.8.2
cmake==3.24.2
wheel==0.41.0
setuptools==68.0.0